refactor: move sdk ecaluation components into one file
